0.3.1 — fix firmware build CI + always-latest firmware delivery
The 0.3.0 release pipeline failed: firmware-build.yml and release.yml
both errored on the Nano Every (megaAVR) target because mcu_id.h
referenced SIGRD (classic-AVR only) and included <avr/eeprom.h> inside
a function body. Wire.requestFrom(addr, count) was also flagged
ambiguous by megaAVR's Wire library.

Result: no GitHub release exists for 0.3.0 and the firmware blobs in
site/public/firmware/ are still v0.2 from March, so the playground at
conduyt.io can't talk to v0.3 firmware (different protocol versions).

Fixes:
- mcu_id.h hoists every vendor header (avr/eeprom.h, avr/boot.h,
  bsp_api.h, pico/unique_id.h, Esp.h, nrf.h) to file scope with
  __has_include guards. Function body just uses the symbols.
  Boot-signature path gated on defined(SIGRD); megaAVR falls through
  to zero-fill or EEPROM-provisioned ID.
- Wire.requestFrom in handleI2CRead/handleI2CReadReg cast explicitly
  to (uint8_t, size_t) to disambiguate against the (int, int) overload.

Always-latest firmware delivery:
- FlashPanel.vue FIRMWARE_BASE now points at
  github.com/virgilvox/conduyt/releases/latest/download. GitHub's
  /latest/download/ URL pattern always resolves to the newest
  published release's asset.
- ESP32 / ESP8266 / Nano-ESP32 manifests rewritten to use absolute
  GitHub release URLs for each part. esp-web-tools follows the
  redirect transparently.
- Removed all 28 committed firmware blobs from site/public/firmware/.
  Only the manifest JSON files remain (those have to be same-origin
  for esp-web-tools to start the multi-file flash). Stops the
  stale-mirror trap where the site ships March firmware while CI
  builds the current one.

Versions bumped 0.3.0 -> 0.3.1 across firmware/library, JS, Python,
Rust, WASM SDKs.

+++ firmware/src/conduyt/ConduytDevice.cpp
@@ -590,7 +590,8 @@ void ConduytDevice::handleI2CRead(const ConduytPacket &pkt) {
     uint8_t respBuf[CONDUYT_PACKET_BUF_SIZE];
     ConduytPayloadWriter w(respBuf, sizeof(respBuf));
 
-    uint8_t received = Wire.requestFrom(addr, count);
+    /* Cast to (uint8_t, size_t) to disambiguate the megaAVR Wire.h overloads. */
+    uint8_t received = (uint8_t)Wire.requestFrom((uint8_t)addr, (size_t)count);
     for (uint8_t i = 0; i < received; i++) {
         w.writeUInt8(Wire.read());
     }
@@ -620,7 +621,8 @@ void ConduytDevice::handleI2CReadReg(const ConduytPacket &pkt) {
     uint8_t respBuf[CONDUYT_PACKET_BUF_SIZE];
     ConduytPayloadWriter w(respBuf, sizeof(respBuf));
 
-    uint8_t received = Wire.requestFrom(addr, count);
+    /* Cast to (uint8_t, size_t) to disambiguate the megaAVR Wire.h overloads. */
+    uint8_t received = (uint8_t)Wire.requestFrom((uint8_t)addr, (size_t)count);
     for (uint8_t i = 0; i < received; i++) {
         w.writeUInt8(Wire.read());
     }
